Generate aten_interned_strings.h automatically
This generates aten_interned_strings.h from `native_functions.yaml`
which is more like how it was originally done. The items deleted from
`interned_strings.h` are duplicates that need to be removed in order
for the code to compile, some of the remaining items may still be out
of date but it is fairly benign even if that's the case.

ghstack-source-id: 784d63a4b5e0c2b8e5612ae46d1dd018dd5e341c
Pull Request resolved: https://github.com/pytorch/pytorch/pull/69407